Split HTMLReader into its own files
