+++ lib/epub_preparer.rb
@@ -3,21 +3,7 @@
 require "zip"
 require "epub/parser"
 require "digest"
-require "open3"
-
-class HTMLReader
-  def initialize(html)
-    @html = html
-  end
-
-  def plain_text
-    Open3.capture2("w3m -dump -T 'application/xhtml+xml'", :stdin_data => html)[0]
-  end
-
-  private
-
-  attr_reader :html
-end
+require "html_reader"
 
 class EPUBPreparer
   def initialize(epub_path, output)
